Factor out CommodityStore as independent struct.
Potentially we can make InternStore as trait,
especially once AccountStore needs to be independent as well.
